Verification and btn Animation
Verification based on the regex for the password field and built in EMAIL format checking. With prompts that something is not correct in the dedicated fields. Also added animation for the button press. Fixed bug that was crushing the app when entered all the credentials and click away to close the keyboard. Fixed UI bug that obsctucted other elements under the focused input field. Changed drawable.