Added Maps
Added maps, in order for them to work on the emulator - download GPS emulator APK and drag and drop on the emulated android deive. APK will be installed, after that open the GPS emulator on the device and choose the location you want to be placed on. Then open FMF and click maps. It should opent the google maps with the harcoded target location which is in the FarmerInfo